Move join tests to a separate package
